+++ packages/eslint-config-smarthr/index.js
@@ -47,7 +47,6 @@ export default [
       'smarthr/a11y-prohibit-input-placeholder': 'error',
       'smarthr/a11y-prohibit-sectioning-content-in-form': 'warn', // TODO: 2024/11にwarn化。問題なければerrorに変更予定
       'smarthr/a11y-prohibit-useless-sectioning-fragment': 'error',
-      'smarthr/a11y-replace-unreadable-symbol': 'error',
       'smarthr/a11y-trigger-has-button': 'error',
       'smarthr/best-practice-for-async-current-target': 'error',
       'smarthr/best-practice-for-button-element': 'error',