+++ packages/eslint-plugin-smarthr/rules/best-practice-for-remote-trigger-dialog/index.js
@@ -1,6 +1,3 @@
-const REGEX_REMOTE_TRIGGER_DIALOG = /RemoteTrigger(Action|Form|Message|Modeless)Dialog$/
-const REGEX_REMOTE_DIALOG_TRIGGER = /RemoteDialogTrigger$/
-
 /**
  * @type {import('@typescript-eslint/utils').TSESLint.RuleModule<''>}
  */
@@ -10,27 +7,17 @@ module.exports = {
     schema: [],
   },
   create(context) {
+    const checker = (node) => {
+      context.report({
+        node,
+        message: `${node.parent.name.name}の${node.name.name}属性には直接文字列を指定してください。
+  - 変数などは利用できません（これは関連するTriggerとDialogを検索しやすくするためです）`,
+      })
+    }
+
     return {
-      JSXOpeningElement: (node) => {
-        const nodeName = node.name.name || '';
-
-        const regexRemoteTriggerDialog = REGEX_REMOTE_TRIGGER_DIALOG.test(nodeName)
-
-        if (regexRemoteTriggerDialog || REGEX_REMOTE_DIALOG_TRIGGER.test(nodeName)) {
-          const attrName = regexRemoteTriggerDialog ? 'id' : 'targetId'
-          const id = node.attributes.find((a) => a.name?.name === attrName)
-
-          if (id && id.value.type !== 'Literal') {
-            context.report({
-              node: id,
-              message: `${nodeName}の${attrName}属性には直接文字列を指定してください。
-  - 変数などは利用できません（これは関連するTriggerとDialogを検索しやすくするためです）
-  - RemoteTriggerActionDialogはループやDropdown内にTriggerが存在する場合に利用してください
-  - ループやDropdown以外にTriggerが設定されている場合、TriggerAndActionDialogを利用してください`,
-            })
-          }
-        }
-      }
+      'JSXOpeningElement[name.name=/RemoteTrigger(Action|Form|Message|Modeless)Dialog$/] JSXAttribute[name.name="id"]:not([value.type="Literal"])': checker,
+      'JSXOpeningElement[name.name=/RemoteDialogTrigger$/] JSXAttribute[name.name="targetId"]:not([value.type="Literal"])': checker,
     }
   },
 }