+++ packages/eslint-plugin-smarthr/rules/prohibit-export-array-type/index.js
@@ -7,18 +7,15 @@ module.exports = {
     schema: [],
   },
   create(context) {
-    const checker = (node) => {
-      if (node.declaration?.typeAnnotation?.type === 'TSArrayType') {
+    return {
+      ':matches(TSTypeReference,ExportNamedDeclaration):matches([declaration.typeAnnotation.type="TSArrayType"],[declaration.typeAnnotation.typeName.name="Array"])': (node) => {
         context.report({
           node,
-          message: '利用する際、配列かどうかわかりにくいため、配列ではない状態でexportしてください',
+          message: `型をexportする際、配列ではなくアイテムの型をexportしてください。
+ - 型を配列でexportすると、その型が配列かどうかを判定するための情報は名称のみになります
+ - 名称から配列かどうかを判定しにくい場合があるため、利用するファイル内で配列として型を設定してください`,
         })
-      }
-    }
-
-    return {
-      ExportDefaultDeclaration: checker,
-      ExportNamedDeclaration: checker,
+      },
     }
   },
 }